Adds threading to chop.seqs
#373 #240

+++ source/commands/degapseqscommand.h
@@ -41,7 +41,7 @@ private:
 	vector<string> outputNames;
 	vector<string> fastaFileNames;
     
-    int createProcesses(string, string);
+    long long createProcesses(string, string);
 	
 };
 #endif
